Getting the conversation started.
Keep in mind that we can always revert back to an older version and discuss what is or isn't working later. So don't be shy to commit your changes and risk a roll back, there isn't much actual risk and the act of starting to contribute is better than sitting on the sidelines asking yourself "is it really okay for me to jump in?" It's okay, jump in.